Handle extension schema attributes
The SCIM specification RFC 7643, chapter 3.3 states:
"Except for the base object schema, the schema extension URI SHALL be used as a
JSON container to distinguish attributes belonging to the extension namespace
from base schema attributes."
An example is given in RFC 7643 figure 5, showing how to properly encode the
attributes belonging to the enterprise user schema extension.

The JSONEncoder now encodes attributes belonging to extension schemas to this
JSON container. The  JSONDecoder is able to handle those JSON container, but
provides backward compatibility.

This addresses the issue #182.