Rebuild charon codebase
The codebase of charon relied up until now on the UserManager.
The codebase has been restyled for making it easier to add additional
resource endpoints.
The UserManager was exchanged for ResourceHandler
The ResourceManager is no longer an interface but an implementation.

The benefits of this can be seen in the impl-project. The resource
handling implementations are now logically seperated from each other.

Another benefit is that the bulk-endpoint can now be implemented with
support for arbitary additional endpoints